[OPTI] islower, isupper (align importance), isalpha, isalnum

+++ libfts.h
@@ -7,7 +7,7 @@ int			ft_isascii(int c);
 int			ft_isdigit(int c);
 int			ft_isupper(int c);
 int			ft_islower(int c);
-int			ft_isalpha(int c);
+int			ft_isalpha(long int c);
 int			ft_isalnum(int c);
 int			ft_isprint(int c);
 int			ft_strlen(char *str);
